Add vrrp_iptables global configuration option
The iptables/ip6tables entries were always added at the end of the
INPUT chain, but for many configurations this is too late in the
processing. This patch allows the chain name to which rules are
added to be specified, and also allows the option of specifying
no rules are to be added.

If a chain name is specifed, it is necessary for that chain to
already exist in the iptables and/or ip6tables config, and for
that chain to be called from an appropriate point in the
ip(6)tables configuration.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>