Use correct MAC address for IPv6 VRRP packets
The IPv6 VRRP packets were using the MAC address of the underlying
interface, rather than the MAC address of the vmac. This commit sets
the correct MAC address for IPv6, and also adds the link-local address
of the underlying interface to the vmac interface, so that VRRP
packets can be sent from the vmac interface, thereby using the VRRP
MAC address.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>