Fix setting nlmsg_len for netlink messages
For netlink messages, nlmsg_len must always be set to an aligned
length. Prior to this commit, nlmsg_len was only being aligned when
a subsequent attribute was added to the list. This was fine if the
length of the last attribute added was an aligned length (which had
always the case), but didn't work if the last attribute added didn't
have an aligned length.

This patch is needed in preparation for adding an attribute which
doesn't have an aligned length.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>