Stop having an IPv6 link-local address added based on VMAC mac address
IPv6 link-local addresses that were based on the virtual MAC address of
the VMAC interface were being added. RFC5798 para 7.4 states that this
is not permitted. It also causes duplicate address detection failure,
since each instance of the virtual router was configuring the same
IPv6 address on the same subnet.

This commit stops the offending link-local address being addied (or removes
it if it can't stop it being added), and since VRRP advertisements must
be sent with the virtual MAC address, but a link-local address for the
interface, if a link-local address from the underlying interface exists,
it is added to the VMAC interface, otherwise the MAC address of the
underlying interface is used to generate a link-local address, which is
then added.

It wasn't until Linux 3.17 that the IFLA_INET6_ADDR_GEN_MODE netlink
message was added, via which one can stop a link-local address being
automatically configured. Therefore, if IFLA_INET6_ADDR_GEN_MODE is not
supported, the only way to ensure that the problematic link-local
address is not added is to remove it after the interface is brought up.
This is not ideal, since there is a small window when the "illegal",
and possibly duplicate, link-local address exists, but I haven't
found any other way of doing it for pre 3.17 kernels.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>