Stop sending unnecessary attributes in netlink messages
When an IPv6 virtual address was deleted, it was being reported
in the log file that preferred lifetime was being set to 0, which
is only relevant when the address is being added. This commit stops
adding the IFA_CACHEINFO attribute when deleting addresses, and
also stops adding other unnecessary attributes.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>