Verify VRRP configuration after all configuration read
There was a lot of duplicated checking in vrrp_parser.c to ensure that configured
parameters were consistent, and also a requirement to configure certain
parameters before others. This checking was incomplete, and also becoming more
and more complex as more configuration options were added.

This commit delays a large part of the checking until after all the configuration
has been read. This removes the need for options to be specified in a certain order
and also for checking in multiple places whether certain combinations are valid.

As a consequence of the delay in checking the configuration, the creation of the
VMAC interfaces is delayed until after the checking.

Although the code appeared to support a mix of IPv4 and IPv6 VIPs on
a vrrp_instance, any VRRP advert sent from a vrrp_instance with mixed IPv4 and IPv6
addresses would have failed at the receive end due to the check that the number
of VIPs in the packet (which could only be EITHER the IPv4 addresses OR the IPv6
addresses) matched the number of VIPs configured locally (which would include BOTH
the IPv4 addresses AND the IPv6 addresses). This commit therefore enforces that the
VIPs, mcast_src_addr/unicast_src_addr and unicast peers are all of the same address
family. If it is desired to have IPv4 and IPv6 VIPs, then one class of addresses
should be configured in virtual_ipaddress_excluded; those addresses will after all
be excluded from the advertisements!

The native_ipv6 keyword becomes redundant, since mixing IPv4 and IPv6 isn't
supported, but it is retained for backwards compatibility.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>