Don't assign VIPs/eVIPs to the default interface
alloc_ipaddress was always setting the interface to DLFT_INT (eth0) if no
dev DEVNAME was specified to a VIP/eVIP/static address. This is fine for a
static address, but doesn't make sense for a VIP or eVIP, since they should
be assigned to the vrrp_instance interface, unless explicitly configured
otherwise.

In fact, it probably doesn't make sense to specify dev DEVNAME for a VIP/eVIP,
since the addresses must be assigned to the vrrp_instance interface.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>