Check for, and handle, '{' at beginning of a block
There was no check for a '{' at the beginning of a configuration block.
This commit is the start of that check, allowing it either at the end
of the line with the keyword, or on a line of its own.

Also, in respect of group and notification_email, for all other configuration
items, the '{' could follow on a line of its own, but for configuration
items using read_value_block the '{' on a line following the keyword
was read as a configuration entry.

The following was valid configuration
group {
VG_1 VG_2
}

and now the following are also valid:
group
{
VG_1 VG_2
}

group { VG_1 VG_2 }

group
{ VG_1 VG_2 }

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>