Ensure an address option has a value
There was no check that the parameter was present after a keyword, so for
example
1.2.3.4 dev
would not have generated an error message, and alloc_ipaddress would have
attempted to read a word after dev, which would either cause a dore dump
or possibly return a parameter from a previous configuration line.

This type of checking probably needs to be added elsewhere too.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>