Make sure sync groups have at least two members
If a sync group was configured with no group {} statement, or if
the group statement had no entries, then keepalived would core dump.
This commit rejects groups with 0 members, and also with 1 member,
since it isn't a group. It also checks that a virtual_instance isn't
configured in more than one sync group, and also that the group
members specified exist.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>