Stop VIPs in same CIDR being deleted, but only when using vmac so far
If an interface has more than one IP address in the same CIDR, when
the "primary" address is deleted, all the secondary addresses are
also deleted, unless /proc/sys/net/ipv4/conf/IFACE/promote_secondaries
is 1. This commit sets the promote_secondaries flag on vmacs.

A later patch will set promote_secondaries on the physical interface
if vmacs are not being used.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>