Allow multiple spaces in quoted strings
The handling of quoted strings saved each word separated between
tokens of '"'. This meant reconstructing a quoted string lost multiple
spaces and was hard work.

Quoted strings are now saved as the whole quoted string, without the
quotes, so retrieval is much simpler. This also allows further keywords
to follow the quoted string, if desired.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>