Ensure vrrp_buffer large enough for largest possible received packet
The allocated receive buffer had size VRRP_PACKET_TEMP_LEN, which
suggests that it wasn't intended as the final solution.

Instead of using a fixed buffer size, the maximum MTU across all the
interfaces is calculated, and the size of the vrrp_buffer allocated
is the maximum MTU size. This guarantees that any VRRP packet received
will fit in the buffer.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>