Remove fixed limit number of VIPs in a VRRP advert
There was an arbirtary limit of VRRP_MAX_VIP (20) VIPs for sending
a VRRP advert. Now that the vrrp_buffer is sized to be able to
receive any packet up to the largest MTU size, we can dynamically
allow as many VIPs as will fit in a packet (which varies depending
on IPv4 or IPv6).

There is also an overhead checking the received addresses in an
advert against the VIPs configure on the instance, but this can now
be mitigated by setting skip_chk_adv_addr on the VRRP instance.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>