Add ability to use libiptc rather than invoking ip(6)tables
Invoking ip(6)tables has a high overhead, since the process has to be
forked and exec'd, and then it has to read the whole ip(6)tables
filter chain before it makes a single update and commits it back.

Using libiptc avoids the overhead of multiple forks/execs, and also
means that multiple entries can be added/deleted to/from the ip(6)tables
configuration in a single update.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>