Fix compile failure with old kernels and libnl3
Issue #215 identified a compile error with pre 3.13 kernels when
libnl3 was installed. This commit adds a test for that situation
and avoids using rtnl_link_inet_[sg]et_conf.

I haven't been able to test this on a re 3.13 kernel, but I have
simulated the scenario and it compiles as expected.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>