Make dependency on libnfnetlink/libnfnetlink.h conditional
Willy Tarreau identified that commit e01618835b introduces a new
unconditional dependency on a libnfnetlink header. Commit e01618835b
was updating configure in line with the changes made in commit
8cd8b4e4 to configure.in.

This commit makes the test for the headers depend on actually using
the netlink libraries, and so matches the needs of
keepalived/include/vrrp_netlink.h.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>