Change outstanding debug flag tests to use bitops helpers
There were two cases where fixed numbers, rather than defined values,
were used to test the debug flag. This commit changes them to use the
defined values and the bitops helpers.

I'm not sure of the logic in scheduler.c of testing for MEM_ERR_DETECT
and DONT_RELEASE_VRRP, but that is what was happening, even prior to
the introduction of bitops.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>