Fix memory leak re some uses of ipaddresstos()
Many of the calls to ipaddresstos() weren't freeing the malloc'd
memory afterwards. This commit changes ipaddresstos() to use a
static buffer, or a buffer passed to ipaddresstos(), thereby avoiding
the use of malloc altogether.

The commit also simplifies dump_iprule() and dump_iproute() by avoiding
the use of intermediate buffers, and reducing the number of malloc's/free's.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>