If VRRP priority is 255 and not nopreempt, configure like state MASTER
If a virtual router has priority 255, the RFCs state that it should preempt
and transition to master mode immediately.

To allow for nopreempt being specified, and also to ensure sync groups can
still work, this commit effectived sets state MASTER if nopremept is not
specified and the priority is 255.

In other words, the following configuration:

priority 255
preempt		# this is the default

will be the same as specifying:

priority 255
preempt		# this is the default
state MASTER

it simply removes the need to specify "state MASTER" if the priority is
255 and nopreempt is not set.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>