Sort out extraneous space and tab characters
The commit removes spaces followed by tabs, trailing spaces and tabs,
and replaces occurrences of 8 spaces within tabs, except where the
spaces and or tabs occur within strings.

This has the benefit that if blocks of code are copied, git does not
complain when running git am on a file produced by git format-patch.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>