Improve and fix finding vmacs left over from previous invocation
When netlink reports a new or existing interface, we can extract
information that allows us to determine if the interface is a macvlan,
and the type (e.g. private). We can then save that in the interface_t
structure, setting the vlan flag, and base ifindex.

When working out the interface name to use for VMAC instances, we can
then check the interfaces which are macvlans to see if any of them
match the vrrp instance in terms of mac address, underlying interface
and inet address family, and if so we can then reuse the macvlan interface.

Commit 9ae463e7f broke the finding of existing interfaces where the
configuration didn't specify the VMAC interface name, and simply created
a new interface. This commits now resolves that.

There is still an issue that if an interface was in MASTER mode when
keepalived terminated, when keepalived restarts it leaves the VIPs and
eVIPS on the interfaces, meaning that keepalived cannot receive VRRP
packets on the interface from the VRRP instance that has taken over, and
it also means that there are duplicate IP addresses on the network.
Another commit will resolve this issue.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>