Ensure ip routes added before rules, and vice versa
If ip rules are added before routes, then it is possible for a
packet to be routed while the routing table is only partially
complete. Adding the rule after the routes ensures that the routing
table won't be processed until it is completely set up.

Likewise, when removing rules and routes, remove the rules first.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>