Add missing reason message for rejected VRRP packet
Issue #255 show a log identifying bogus VRRP received, but there
was no reason shown for the rejection. The only instance I can find
for this is if vrr->family is neither AF_INET or AF_INET6, which I
think must be a bug in the code parsing and setting up VRRP instances.

This commit just adds a log message to be explicit about why the packet
is rejected, and also reports the value of vrrp->family.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>