Report if vrrp or checker process abnormally terminates
Issue #159 identifies that keepalived doesn't always remove the VIPs
when stopping. The only instance I hace seen of this is if the vrrp
processes terminates due to a segmentation fault while it is
terminating.

This commit makes the parent process log if a child process terminates
other than with a successful exit code. In particular it requests that
an issue be raised if the process terminates with a segmentation fault.

In order to be able to report the child process termination, the parent
process now has to wait when a SIGTERM has been sent to its children in
order to receive the exit status of the children.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>