Add support for identifying git commit version at runtime.
It is useful to be able to know exactly what version of the code is
being executed, especially if someone is reporting an issue.

If the version of keepalived is not a tagged version in git, the
tag and commit will be reported by keepalived -v and in the startup
and shutdown log messages.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>