Add option to increase child process priorities and make non swappable
Issue #211 (and others) have identified that under heavy load, backup
instances can stop seeing the VRRP adverts from the master, and the
backup then promotes itself to master, causing two masters.

This commit adds the option to increase (or decrease) the priority of
the vrrp and checker processes, and also to lock the processes into
memory.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>