Don't restore original signal state when reloading vrrp config
Issue #209 identified that sending two SIGHUP signals to the vrrp
process in quick succession could cause the vrrp process to terminate.

This commit removes the restoring of the default signal state and then
setting it back to the running signal state, so that the signal handling
disposition if not changed at all during a reload.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>