Don't overwrite real interface MAC address with VMAC MAC address
When a VMAC was being created, the MAC address of the VMAC was
being copied to the MAC address of the underlying interface in the
interface_t structure.

The netlink reflector sets up the MAC address of the new VMAC
interface, so there is no need to copy a MAC address at all.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>