Make lvs_sync_daemon global config rather than vrrp specific
Issue #237, and various emails over the years have indicated
confusion due to lvs_sync_daemon being configured on the vrrp
instances, and hence suggesting that it could be configured
on multiple vrrp instances.

This commit changes how the lvs_sync_daemon is configured, so
that it is part of the global configuration, and not configured
within a specific vrrp_instance. In order not to break any existing
configurations, configuring lvs_sync_daemon within a vrrp instance
is still supported, but logged as deprecated. In addition, if
lvs_sync_daemon is configured on more than one vrrp_instance, an
error is logged.

In making this change, it was also discovered that when reloading
keepalived with a new configuration, if the lvs_sync_daemon was
moved onto a different vrrp instance, then depending on the order
of the vrrp instances, the lvs sync daemon could be started on the
new instance (which would fail) but then subsequently stopped on the
old instance, leaving the sync daemon not running.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>