Send second set of GARP messages afer receiving lower prio advert
When a VRRP instance transitions to master state, if garp_master_delay
is non-zero, a second set of garp_master_repeat messages is sent after
garp_master_delay seconds (unless 0). However, if a lower priority advert
is received, keepalived didn't send a second set. This commit sends a
second set if a second set would have been sent after transition to master.

Thanks to flicker581 for identifying the issue and providing a patch.

This resolves issue #277.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>