Add support for iprule and iproute table names
Pull request #153 added support for named tables in adding/removing
iprules and iproutes.

This commit updates that pull request to use standard keepalived
functionality, and also to support named tables for iprules. In
addition, it supports table numbers > 255.

When reading a configuration, if there is a static or virtual rule
or route entry, and if the table specified is not numeric, then the
table names/numbers are read from the /etc/iproutes/rt_tables file.
The table names are discarded after the configuration has been read,
and if the configuration is reloaded, then the table names, if needed,
are read again. If there are no rule or route entries, or if they all
specify tables by number, then the table names file is not read.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>