Restore sync-state after reload
Currently the sync state is rebuilt from the member states after
config reload. This changes now reloads the previous sync state
after reload, and then pushes this back to the group members. If a
new group member is added during the reload, then the new group
will accept the sync group state. If a group member is removed
during a reload, then a special case will be executed to force
the sync-group state to BACKUP. This is required so that an
alternative backup peer for the removed group is given an
opportunity to take over the gateway.

Fixes issue #285