Report and exit if configuration file(s) not found/readable
The configuration file is treated as a pattern, and processed
using glob(). If there is no matching file, then it wasn't reading
any file, and keepalived was running with no configuration.

This patch adds a specific check that there is at least one matching
file, and also checks that all the configuration files are readable,
otherwise it reports an error and terminates.

This resolves issue #304

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>