Add options -m and -M for producing core dumps
Many systems won't produce core dumps by default. The -m option
sets the hard and soft RLIMIT_CORE values to unlimited, thereby
allowing core dumps to be produced.

Some systems set /proc/sys/kernel/core_pattern so that a core file
is not produced, but the core image is passed to another process.
The -M option overrides this so that a core file is produced, and
it restores the previous setting on termination of the parent process,
unless it was the parent process that abnormally terminated.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>