Add aggregation of interfaces for throttling ARPs/NAs
Commit 42dbac9 added support for throttling gratuitous ARPs and
unsolicited NAs, but dealt with it on a (physical) interface by
interface basis.

This commit adds support for aggregating interfaces together, so
that if multiple interfaces are connected to the same physical switch
and the switch is limited as a whole on the rate of gratuitous ARPs/
unsolicited NAs it can process, the interfaces can be grouped together
so that the limit specified is applied across them as a whole.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>