In free_interface_queue, don't check LIST_ISEMPTY before freeing
LIST_ISEMPTY() returns false if the queue exists, but there are no
entries on the queue. If free_list is only called is !(LIST_ISEMPTY)
then the queue struct itself will not be freed.

This commit removes the check of LIST_ISEMPTY, to ensure that the
queue structure itself will always be freed.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>