Update master_adver_int when receive higher priority advert when master
If VRRPv3 is being used, and a higher priority advert is received when
in master mode, the master_adver_int needs to be updated when transitioning
backup mode. If this isn't done, and our advert interval is less than a third
of the new masters, we will time out and re-enter master mode, send an advert
to which the other master will resond with a higher priority advert, causing
us to go back into backup mode, until our timer expires again, and this will
continue indefinitely.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>