Fix receiving advert from address owner when in fault state
If a vrrp instance was in fault state, and it then receives an
advert from the address owner (advert priority == 255), the
instance would transition to master mode. The check was:
      } else if (vrrp->effective_priority > hd->priority ||
                  hd->priority == VRRP_PRIO_OWNER) {

Clearly, if the advert is FROM the address owner, then this
instance should be transitioning to backup state, and not master
state.

A reasonable check would have been:
      } else if (vrrp->effective_priority > hd->priority ||
                  vrrp->effective_priority == VRRP_PRIO_OWNER) {

but the second part of the test is superfluous since if
vrrp->effective_priority == VRRP_PRIO_OWNER, then hd->priority
should not be VRRP_PRIO_OWNER, and hence we will have
vrrp->effective_priority > hd->priority

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>