Fix IPv4 vIP removal when addr matches pre-existing interface addr
For IPv4 vIPs keepalived adds a /32 to the underlying interface. If
this address matches an address already configured, e.g. a /24, when
this vIP is eventually removed due to a configuration change or
keepalived shutdown, the original address matching the vIP, outside
of keepalived's control, is removed instead. This behaviour is
incorrect. The /32 added by keepalived should be the address being
removed. Keepalived should not be touching any addresses it does not
create.

This commit re-instates d80c171, the contents of which were removed
in ef73279. The IFA_ADDRESS attribute is included in the netlink
delete address message so that the /32 IPv4 address keepalived adds
is removed instead of the pre-existing interface address.