Fix transition to backup when receive equal priority advert from higher address
When a vrrp instance in master mode received an advert from another master that
had equal priority, it wasn't comparing the addresses to determine whether it
should treat the advert as higher priority, and hence the instance should
fall back into backup state.

When checking whether the advert is from a lower priority master, it now checks
if the priorities are equal and then compares the addresses.

This commit resolves issue #342.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>