Make malloc/free diagnostics a separate configure option
The commit adds the configure --enable-mem-check option which
allows the MALLOC/FREE diagnostics to be enabled without
the --enable-debug option. This means that the mem-check
diagnostics can be used when running keepalived in it's normal mode
with forking children for vrrp and checkers.

The mem-check diagnostics are written to
/tmp/Keepalived_{,vrrp,healthcheckers}_mem.PID.log

The --mem-check-log configure option enables command line option
-L which also writes zalloc/free details to the syslog.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>