Fix segfault or infinite loop in thread_child_handler() after reloading
When the checker and vrrp child processes start up, memory for a
thread_master_t is malloc'd and saved in master. Subsequently,
launch_scheduler() is called, and that sets the parameter to be passed
to the SIGCHLD handler - thread_child_handler() to the value of master,
pointing to a thread_master_t.

If keepalived is signalled to reload, the child processes free all
malloc'd memory, and a new thread_master_t is malloc'd and saved in
master. If this is not the same address as the previous thread_master_t,
then the value being passed to the SIGCHLD handler is a pointer to the
old thread_master_t, whereas everything else is using the new thread_master_t.
If the memory used for the old thread_master_t is then returned in a subsequent
malloc() call, a subsequent SIGCHLD will invoke thread_child_handler() with
a pointer to memory that has now been overwritten for some other purpose, hence
causing either a segfault or an infinite loop.

A further consequence is that new child processes will be added to the new
thread_master_t, but when thread_child_hander() is called after a child
terminates, it won't find the child since it is still looking at the old
thread_master_t.

This commit modifies the behaviour of a reload by not releasing the old
thread_master_t and then malloc'ing a new one, but rather it just reinitialises
the original thread_master_t and continues using it.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>