Add back real server when return from failure with HTTP_CHECK
If status_code wasn't specified for a url entry in the configuration
then a real server would never be returned to service following a
failure.

The commit makes keepalived return a real server to service if no
status_code is specified if the HTTP status code returned from the
service is a success code (i.e. 2xx).

This should resolve issue #373

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>