Ensure old VIPs removed after reload
Issue #382 identified that if VIPs were removed from a configuration
and the configuration reloaded, then the VIPs that had been removed from
the configuration were left configured on the interfaces.

The issue only occurred if accept mode was set on the vrrp instance, and
was due to a wrong condition being tested.

This commit changes the condition being tested from
(!old_vrrp->iptable_rules_set) to (!old_vrrp->vipset).

This should resolve issue #382.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>