Add network namespace support.
This allows multiple instances of keepalived to be run on a single
system. The instances can communicate with each other as though they
are running in separate systems, but they are also isolated from
each other for all other purposes.

See keepalived/core/namespaces.c for some example configurations and
use cases.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>