Add polling of netlink messages when entering master state
If a large number of vrrp instances enter master state simultaneously
the netlink socket can run out of buffers, since the netlink socket
isn't read sufficiently frequently. Adding a poll of the netlink socket
after the VIPs/eVIPs are added ensures that the netlink messages are read
when the become available.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>