Add global default_interface keyword
default_interfaces sets the default interface to use for static
ipaddresses. If the system does not have an eth0, or one wants to
use a different interface for several static ipaddresses, this makes
the configuration simpler. It also has the potential to reduces
changes required if transferring the configuration to another system.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>