Add support for running multiple instances of keepalived
Using network namespaces allows multiple instances of keepalived to run
concurrently, in different namespaces, without any collision of the pid
files.

This patch adds the concept of a keepalived instance name, which is then
use in the pidfile name, so that multiple instances of keepalived can run
in the same namespace without pid file name collisions.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>