Check rx packet priority >= our priority to terminate preempt delay
If a VRRP instance is in backup mode it should treat receiving an
equal priority message the same as a higher priority message, but
when a preempt delay timer was running, it was only doing this for
higher priority message. The consequence of this was that two
equal instances both with preempt delay could end up flip-flopping.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>