Become master as soon as send first advert
After sending the first advert after deciding to become master
we were waiting a further advert interval before configuring the VIPs
etc on the interface. This meant that although we had told the network
we were the master, for 1 advert interval we weren't doing so, and so
the virtual router would be down for that period.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>