Ignore lower priority adverts when backup
Although sending an advert when in backup mode if a lower priority
advert is received should make the transition to master occur sooner,
doing so is not compliant with the RFC, which requires that lower
priority adverts are simply discarded. Doing so doesn't stop the
transition to master, but it does mean that it will take place in an
orderly fashion, albeit taking 3 advert intervals + skew time.

The problem with sending an advert immediately after receiving a lower
priority advert is that other higher priority routers will all do so at
the same time, and confusion will ensue as to who is master. It is necessary
to build in the skew time in order to have an orderly transition.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>