Improve finding a timed out vrrp thread
Returning the vrid, to then search for the vrrp_t * by vrid was
not optimal, when the vrid was determined from the vrrp_t * anyway.
This also adds a check when searching through the vrrp_index_fd[]
that the fd_in of the vrrp_t * found matches the fd_in that we want.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>