Fix removing tracked scripts from sync group members
If a vrrp instance belongs to a sync group, there shouldn't be any
weighted track scripts run for the instance. This was being checked,
and if such a weighted track script was found, the use count of the
script was decremented. If that track script was used on non
sync-group interfaces, the inuse count would not be decremented to 0,
and so the script would not be marked as disabled. As a consequence,
the weight of the script would still alter the priority of the vrrp
instance.

This commit adds removing the script from the tracked_script list on
the vrrp instance if it is a member of a sync group and the script
is weighted thereby stopping the script changing the priority of the
vrrp instance.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>