Make vrrp enter fault state at startup if interfaces down
Previously the fault state of an interface was only checked when a send or
receive timeout expired. If the interface went down just after the last advert
was received, and came back up just before the ms_down_timer expired, then
the vrrp instance would transition to master since it hadn't received any
adverts but also hadn't detected that the interface was down. The same problem
occurred if an interface was down when keepalived started, but subsequently came
up.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>