Don't reflect interface state flags from real interface to vmac
keepalived was reflecting the interface state flags (e.g. UP) from
the underlying real interface to the vmac interface, with the consequence
that if the vmac interface was down (e.g. a user did 'ip link set vrrp.1 down'
for example), then keepalived not receive anything on the vmac interface,
but would also not detect that it was down. As a consequence, even though
the interface was down, the vrrp instance would become, or remain, master.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>