Implement not accepting packets to VIPs for VRRPv2
VRRPv3 was blocking packets sent to VIPs (and eVIPS), unless the
vrrp instance was the address owner, or accept mode was set.

RFC3768, for VRRPv2, specifies that packets addressed to the VIPs
should not be accepted, unless the router is the address owner. This
commit implements not accepting the packets when running VRRPv2, but
also adds an accept mode, to allow the packets to be received, unless
strict mode is set.

This commit alters the default behaviour of keepalived, by, by default,
not accepting packets to VIPs when using VRRPv2.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>