Make report_child_status() check for vrrp and checker child processes
report_child_status() checks for exit status KEEPALIVED_EXIT_FATAL
and KEEPALIVED_EXIT_CONFIG, but these are only relevant for the vrrp
and checker child processes, and not for track scripts etc. This commit
adds a check that the terminating process is the vrrp or checker process
before checking those exit statuses.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>