Add no_accept mode for VRRPv2 and standardise VRRPv3 with it
RFC3768, for VRRPv2, specifies that packets addressed to the VIPs
should not be accepted, unless the router is the address owner.

This commit implements not accepting the packets when running VRRPv2,
but only if no_accept is specified, or running in strict mode. The reason
for not making no_accept the default (which would confirm to the RFCs) is
that if running IPVS, or any other service on top of the VIPs, we need to
be able to accept the packets, and requiring everyone to specify accept
in that case would not be reasonable.

Prior to this commit, VRRPv3 was blocking packets sent to VIPs (and eVIPS),
unless the vrrp instance was the address owner, or accept mode was set. This
commit changes the default behaviour for VRRPv3 to make it consistent with
VRRPv2 (i.e. either strict mode or no_accept needs to be specified to be
conformant with RFC5978).

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>