Fix address ranges for virtual server groups
The handling of address ranges was only written for IPv4 addresses, and
only worked on little endian systems.

This commit enables IPv6 address ranges to work, and also should now
work on big endian systems (but I don't have access to a big endian system
to test it). Validation is added to ensure that the end of the range is after
the start of the range, and that the value of the range end does not exceed
255 (for IPv4) of ffff (for IPv6).

There is also some optimisation of the code, so that netmask is not set (since
it isn't used by the kernel), and the port is set once only, before the loop
through the addresses.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>